Table formatting
Okay so a feature I think we really need are condensing tables. We need users to be able to  tap a table to expand it to save space, because otherwise I don't think I can include tables of ideas which is IMO the best way to present ideas to DMs for world building. The idea is to generate clearly skippable content that is easy to reference for ideas.

Also it is super important that we can export content elsewhere for reference to save space. We need a repository for both individual content that pairs with articles I can link two as well as a general large source that contains and updates as I produce more.